requirements.txt
# Combined dependencies for Python 3.11.9
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
transformers==4.44.2
torch==2.4.1
langchain-groq==0.2.0
python-dotenv==1.0.1
gtts==2.5.3
newsapi-python==0.2.7
pydantic==2.9.2
streamlit==1.38.0
langchain==0.3.1
duckduckgo-search==7.5.2
langchain-community==0.3.1
google-search-results==2.4.2
googletrans==4.0.0-rc1